requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
ftputil==3.4
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
pony==0.7.13
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.4.1
selenium==3.141.0
six==1.14.0
suds-jurko==0.6
urllib3==1.25.8
wcwidth==0.1.8
